Login with google done and tested. Add meetings to calendar api done but not tested.

+++ public/js/app.js
@@ -1,5 +1,6 @@
 import { api, goToHome, getFormattedDateString } from "./helpers/helpers.js";
 import * as FilterFiles from "./filter/filter-files.js";
+import { googleSignOut } from "./google/googleOAuth.js";
 //const flatpickr = require("flatpickr");
 
 api.defaults.headers.common['token'] = localStorage.getItem('token')
@@ -188,8 +189,12 @@ const getDateObjectFromDateTimeInputs = () => {
 
 //#region event callbacks
 function signOut() {
-  localStorage.clear()
-  goToHome()
+  if(localStorage.getItem('googleSign')) {
+    googleSignOut()
+  } else {
+    localStorage.clear()
+    goToHome()
+  }
 }
 function updateFieldIfNecessary(e, clinicInputs) {
   //console.log(e)
